requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
billiard==4.2.1
blis==0.7.11
catalogue==2.0.10
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cryptography==44.0.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
defusedxml==0.7.1
Django==5.1
django-celery-results==2.5.1
django-environ==0.12.0
docx2txt==0.8
easyocr==1.7.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gunicorn==23.0.0
idna==3.7
imageio==2.34.2
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.5
kmodes==0.12.2
kombu==5.4.2
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
Levenshtein==0.26.1
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
meson==1.5.1
mpmath==1.3.0
murmurhash==1.0.10
mysql-connector-python==9.1.0
mysqlclient==2.2.4
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
olefile==0.47
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
pdfminer.six==20240706
pillow==10.4.0
plotly==5.23.0
plotly-express==0.4.1
preshed==3.0.9
prompt_toolkit==3.0.47
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyclipper==1.3.0.post5
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyrsistent==0.20.0
pytesseract==0.3.10
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-Levenshtein==0.26.1
pytz==2024.1
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.21.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sortedcontainers==2.4.0
spacy==3.7.6
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.2
style==1.1.0
sympy==1.13.1
tenacity==9.0.0
tesseract==0.1.3
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.7.24
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
update==0.0.1
urllib3==2.2.2
use==0.1.0
utility==1.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
whitenoise==6.9.0
wrapt==1.16.0
